Create README.md
The main purpose of this site is to order and distribute faculties and their students within a university. 
The user's first interaction with the site is through the Login page. After creating the account, the user is redirected to the main page of the site. From this menu, we can access both the list of faculties and students, as well as the contact details of the university. On the page intended for faculties, we will find the list of those already registered, the departments and the deans, also having the possibility to carry out other actions such as updating or deleting them. Moreover, students have the possibility to submit their applications to a specific faculty directly through the website.